+ AbstractAlgoInput and Result introduced with helper for options. + Some bugfixes in Jail.ag and GMatch.ag + Random graph and fa algos are on the way. + Unable exception introduces for random algorithms. + Semantics of IDirectedGraphPath ajusted on emtpy paths.
git-svn-id: http://svn.chefbe.net/be.uclouvain.jail/trunk@57 bb353e37-10b9-4b19-88b1-c35c3065af8f

+++ src/algo/be/uclouvain/jail/algo/graph/copy/match/functions/ChooseFunction.java
@@ -23,7 +23,7 @@ public class ChooseFunction implements IGMatchFunction {
 	public Object execute(Object...args) {
 		// randomly generates an integer
 		if (args == null || args.length==0) {
-			return new Integer(r.nextInt());
+			return new Double(r.nextDouble());
 		}
 		
 		// randomly chooses a value from args
